% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{scan_data}
\alias{scan_data}
\title{Load columnar, in-memory data}
\usage{
scan_data(path, is_dir = FALSE, pattern = "*")
}
\arguments{
\item{path}{String or path-like object, points to raw input file(s)}

\item{is_dir}{Boolean, signal for function to recursively apply scan to all files}

\item{pattern}{String, regular expression for files to scan from directory}
}
\value{
Arrow dataframe
}
\description{
Load columnar, in-memory data
}
\examples{
# scan single file
scan_data(file.path(path="~/path/to/file.ext"))
# scan multiple files at directory
scan_data(file.path(path="~/path/to/files/"), is_dir=TRUE)
# scan multiple files at directory of specific pattern
scan_data(file.path(path="~/path/to/files/"), is_dir=TRUE, pattern="*.txt")
}
